requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-js-asset==1.1.0
django-mptt==0.9.0
django-storages==1.6.6
dropbox==8.9.0
gunicorn==19.8.1
idna==2.6
Pillow==5.1.0
pipenv==11.10.1
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
